+++ lib/main.dart
@@ -2,6 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:easy_localization_loader/easy_localization_loader.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 import 'app.dart';
@@ -16,12 +17,14 @@ Future<void> main() async {
     anonKey: dotenv.env[Constants.supabasePublicDefaultKey] ?? '',
   );
   runApp(
-    EasyLocalization(
-      supportedLocales: const [Locale('en'), Locale('zh', 'TW')],
-      path: 'assets/translations/strings.csv',
-      assetLoader: CsvAssetLoader(),
-      fallbackLocale: const Locale('en'),
-      child: const App(),
+    ProviderScope(
+      child: EasyLocalization(
+        supportedLocales: const [Locale('en'), Locale('zh', 'TW')],
+        path: 'assets/translations/strings.csv',
+        assetLoader: CsvAssetLoader(),
+        fallbackLocale: const Locale('en'),
+        child: const App(),
+      ),
     ),
   );
 }