+++ lib/shared/models/auth_model.dart
@@ -1,0 +1,6 @@
+class UserData {
+  final String email;
+  final String uuid;
+
+  UserData({required this.email, required this.uuid});
+}