+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## v0.3.0
+更新內容:
+- 支持「郵箱密碼」登入
+- 支持「郵箱密碼」註冊
+
+Updates:
+- Support sign in with email and password
+- Support sign up with email and password
+
+## v0.2.0
+更新內容:
+- 新增「登入介面」
+- 新增語言系統（支持繁體中文、英文）
+
+Updates:
+- Add Sign up page
+- Add Localization system (support en, zh_TW)
+
 ## v0.1.3
 更新內容:
 - 新增「更新檢查」